add backgrounds to watch script separate into modules
