+++ src/scripts/watch/constants.mjs
@@ -1,0 +1,4 @@
+export const optionsBase = {
+  awaitWriteFinish: true,
+  cwd: process.cwd()
+};
